Moved styles into separate file
